requirements-common.txt
pandas
requests
lxml
beautifulsoup4
pytz
rich
jsonschema<4.18.0
pymarc
reportlab
python-barcode
qrcode
google-auth
pytest
pytest-asyncio